+++ src/api/schemas/review-store.shop.schema.ts
@@ -29,7 +29,7 @@ export const reviewStoreShopApiExtension = gql`
   }
   extend type Query {
     "Get the average of review store"
-    avgReviewStore: Int
+    avgReviewStore: Float
 
     "Get the review store of the current customer"
     myReviewStore: ReviewStore
